pyproject.toml
[project]
name = "minebot"
version = "1.1.0"
description = "A modern, all-in-one Discord bot with a Minecraft connector and a ton of useful features."
requires-python = ">=3.10"
dependencies = [
    "aiomysql==0.2.0",
    "aiosqlite==0.21.0",
    "asyncpg==0.30.0",
    "hikari-lightbulb==3.0.1",
    "hikari-toolbox==0.1.7",
    "hikari[speedups]==2.3.3",
    "pydantic==2.11.7",
    "pydantic-extra-types==2.10.5",
    "sqlalchemy==2.0.41",
    "websockets==15.0.1",
    "uvloop==0.21.0; sys_platform != 'win32'",
    "cryptography>=45.0.3",
    "hikari-chat-exporter>=0.0.3",
    "pygithub>=2.6.1",
]
